A couple of corrections.  One place had the word "commas" when "semicolons" was clearly intended.
Another is removing the remark that Javascript's implicit access of global variables is nondeterministic. It is only nondeterministic if there are two places that access the same variables that are called in response to events. If the variable is first set as a global in Move then the same Move program is equally nondeterministic.